+++ ecorise-react-app/src/Payment/index.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import PaymentForm from './PaymentForm';
+import PaymentForm from '.';
 
 
 jest.mock('../hooks/useStkPush', () => ({
