+++ ecorise-react-app/src/hooks/useStkPush.js
@@ -1,5 +1,4 @@
 
-
 import { useState } from 'react';
 import { stkPush } from '../utils/paymentApi';
